cmd/node: use shared filelock by default (#4373)
originally i added file lock for a specific data directory, as it seemed nicer that we can run multiple instances of go-spacemesh. however there are ongoing reports that people somehow run more than one, and submit different atxs, etc.

it would be very stupid if we will have malfeasence proofs because of that. so i would rather request people to configure separate file locks if they wish to run many instances on same computer, to do so there is an option `--filelock`, that defaults to `/tmp/spacemesh.lock`