mesh: rework processing of the consensus results (#4383)
most of this change is about making consensus results processing to be more straightforward. and it also removes fetching of the blocks referenced in ballots.

tortoise outputs latest consensus results with computed opinion as in example:
```json
[
  {
    "layer": 24,
    "opinion": "55bfc8b672",
    "verified": true,
    "blocks": [
      {
        "id": "59e7b07d06",
        "layer": 24,
        "height": 23563,
        "valid": true,
        "invalid": false,
        "hare": true,
        "data": true
      }
    ]
  },
  {
    "layer": 25,
    "opinion": "1a5fe28f37",
    "verified": false,
    "blocks": [
      {
        "id": "e2cda88192",
        "layer": 25,
        "height": 23563,
        "valid": false,
        "invalid": false,
        "hare": true,
        "data": true
      }
    ]
  }
]
```

mesh is supposed to take this result and ensure that the vm state (and the rest of db) is consistent with results computed by tortoise. 

tortoise tracks when opinion has changed, and therefore it can provide results that start before currently applied layer. in such case mesh goes over layers under the question and finds which one it needs to revert.

after that mesh applies all layers consistently with tortoise results. with the nuance that we want to apply hare within hdist, unless block upvoted by hare crossed negative threshold. see layer 25 from the json above.