Drop p2p connections on ValidationReject (#4374)
## Motivation
We don't want to stay connected to peers that we know are exhibiting invalid behaviour.
This is a step towards completing #4088

## Changes
Change the signature of `pubsub.GossipHandler` to return an error instead of a `pubsub.ValidationResult` this makes the interface used by both sync handlers and pubsub handlers consistent, which allows for the use of a single decorator that will drop a peer if the handler returns a specific error. The decorator (`pubsub.DropPeerOnError`) is applied to all pubsub handlers and sync handlers.

Added a test to verify disconnection on pubsub handlers, to support this I exposed the `p2p.Host` from `node.App`

## Test Plan
The test shows that this is working. A test or tests to verify this behaviour for sync handlers has yet to be added

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)