logging: revisit verbosity (#4442)
there are multiple logs that print duplicate information on progress, this significantly decreases readability.
for example optimistic execution printed 3-4 logs with similar info, which is clearly useful only for debug. if thats necessary developer should raise that module to debug level.

in some other cases metric is a cheaper and often better alternative, e.g a count of some object or computed value.