tortoise: tracer to debug execution interactively (#4447)
tortoise provides EnableTracer option, if enabled it will dump to stderr all inputs and outputs from tortoise consensus using zap library.

this logs will be available in cloud and can be downloaded either with grafana or logcli. if downloaded from grafana, next post processing script is useful:
> jq -c '.[].line | fromjson | {"t": .t, "o": .o}' ./EXAMPLE.json

note that traces are verbose and should be decreased in size, as an example i added tortoise/data/failed_nodes.json from one of the nodes.

once trace is collected it can be used as a regression test, and also for interactive debugging. for interactive debugging:
1. install delve https://github.com/go-delve/delve/tree/master/Documentation/installation
2. run delve with selected trace `dlv exec ./trace -- -breakpoint -level=debug ./tortoise/data/failed_nodes.json`
3. continue will stop after every step, debug log will also print what happened during that step
```
> (dlv) continue
2023-05-30T11:42:03.864+0200	DEBUG	trace	using trace	{"path": "./tortoise/data/failed_nodes.json", "name": ""}
> [hardcoded-breakpoint] github.com/spacemeshos/go-spacemesh/tortoise.RunTrace() ./tortoise/tracer.go:101 (hits total:0) (PC: 0xdf038f)
Warning: debugging optimized function
    96:				return err
    97:			}
    98:			if err := ev.Run(runner); err != nil {
    99:				return err
   100:			}
=> 101:			if breakpoint != nil {
   102:				breakpoint()
   103:			}
   104:		}
   105:	}
   106:	
```