systest: partition tests (#4450)
closes: https://github.com/spacemeshos/go-spacemesh/issues/4402

the result of the partition could be that some block crosses local threshold, but doesn't cross the global threshold. lets say in total weight the fractions would be 55/45. 55 voting in support, 45 voting against the block. once local threshold is crossed  55-side will be always voting in support, 45-side will be seeing that block crosses local threshold, but not global, but won't have data to vote for the block. as such it won't be possible to reach global threshold indefinitely, as it will always remain at ~10 of total weight, as 45-side will never ask for the block.

the change is to force 45-side to ask for the block even if it crossed local threshold.

other changes:
- make traces shorter by removing all unnecessary data (new types to communicate with tortoise) and encoding byte arrays (ids, node) with base64
- enable lfs for tortoise/data/
- collect traces from multiple ci tests for regression testing