Adjust structure of protocol and handler
The broker needs to know from the handler whether it should relay a
message to the network. Having that communication happen through a
channel seems over complicated, so instead the broker will call handle
directly and the handler and the protocol will use a mutex to
synchronise their actions. Neither the handler or protocol contain any
blocking code (except for the lock) and the code is not performing
costly transactions so this shouldn't pose a problem to the broker by
causing the broker to wait.