Thinking about the path to take
I think we can run everything (including early messages) through the
handler and then construct the actual protocol bit later, because really
the handler is just a bunch of accumulators.

The set Si can be provided to the protocol when we start to run it, it
doesn't need to be given at construction time.