Wip
have decided that we process early messages fully through the hare3 handler
and that that means that we need to build the hare3 handler in the
broker in response to early messages.

That means that we want to return the gradecaster and thresh graded
gossiper upon consensus process registration so that those can be built
into the consnesus process.

This may not be an ideal final design since the contents of hare/hare.go
are looking less and less useful, but for now we'll do this to get the
broker integration out of the way.

It feels to me that ultimately the hare and broker will merge to become
one thing and a lot of the hare functionality will be repackaged into
smaller mor targetted services.

Next step right now is to split the handler object defined in the broker
to be a message store and a separate handler.