tx: get rid of sleeps in tests (#4597)
## Motivation
In #4575 sleeps were added to tx tests to ensure that they pass on windows. This fixes the code to not require the sleeps any more.

## Changes
TX are now sorted in the priority queue according to the following:

- largest fee
- if equal earliest receive time
- (new) if also equal transaction ID

on *nix systems it is near impossible for two transactions to be received on at the same time (within 2 ns). On windows however for two timestamps to be considered equal they can be up to 15 ms apart. Tis might cause the order of transactions to randomly change in the pool, that's why an additional check was added to determine the priority.

## Test Plan
<!-- Please specify how these changes were tested 
(e.g. unit tests, manual testing, etc.) -->

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)