txs: fix flaky test on windows (#4602)
## Motivation
This fixes the flaky test `TestSelectProposalTXs` on windows.

## Changes
`SelectProposalTXs` is supposed to return a random collection of transactions from the mempool. To generate this random set it uses `time.Now().UnixNano()` as a seed. On windows this will return the same integer if called twice within less than ~ 15 ms.

The test is fast enough that between the first and second call to `SelectProposalTXs` no new timestamp will be available. 

The code was updated to not seed the PRNG on every call, but rather only once.

## Test Plan
n/a

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)