Fix copy paste mistake in function comment (#4666)
This commit writes a comment for `handleMaliciousIDsReq` that matches its functionality. I tried to understand what the code does, but not sure if `all malicious nodes` is the right description (is it only some of them?).

Background: I'm doing a research study on function comments that may be improvable. You can find more info and a short and optional anonymous survey [here](https://forms.office.com/e/yHv4PRXM2i).

Thanks for reviewing!