api, events: cache last 100 admin events in memory (#4669)
grpc subscriber might be slow to join or disconnect due to flakes and other reasons, we don't want it to miss
important events because of that.

added simple ring buffer, and used it to cache last 100 user events. they are copied (pointers) and are sent to the newly created stream.