fetch: requests got lost when there are no peers (#4849)
requests are not restored back to the queue when there were no peers to dispatch to.

- restore requests when there are no peers
- push mesh to still make progress even when there are missing blocks
- increase frequency of state sync

i manually test against mainnet, artificially causing GetPeers() to return zero peers.
mesh was making very slow progress because it aborts as soon as it finds a missing block, not executing the next block even when it's available.