Ensure that App closes started channel on startup error (#4869)
## Motivation
Closes #4868 

## Changes
This simply moves non blocking app startup operations into a new method allowing for a deferred call to close the started channel. This ensures that started will always be closed before Start returns and thus prevents getting stuck waiting for the `App` to start. There were also a couple of services that were deferring their shutdown to the end of Start, they have now been added as fields of` node.App` and are shutdown in `stopServices` which keeps the shutdown approach consistent.
## Test Plan
Current tests should pass

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)