Simplify scrypt in systests and reduce batch size (#4995)
## Motivation
Nodes in systests sporadically hang during POST init. The last message is:
```
initialization: continue looking for a nonce	{"startPosition": 256, "batchSize": 1048576}
```

It hangs because the batch size is big and init is performed on CPU (it should eventually finish, it just takes veeery long). This fix changes the batch size to match `labelsPerUnit` for the fastnet preset (which systests use).

## Changes
- reduce batch size to 128 in the fastnet preset,
- reduce scrypt difficulty to the lowest possible N = 2 in the fastnet preset.

## Test Plan
Existing tests pass