Fix flaky TestAccountMeshDataStream_comprehensive (#5075)
## Motivation
Closes #4171 

Occasionally the test will fail with an incorrect assertion of an error. The stream receive can fail either because the deadline was exceeded or due to bad timing for another (random) reason. Since it doesn't matter which return code is received, just that the stream ended I changed the assertion to also allow `code.Unkown`.

## Changes
- change assertion to allow `codes.DeadlineExceeded` and `codes.Unkown` instead of just the former.
- Additionally get rid of the global GRPC config. This will allow to run GRPC service tests in parallel (which will be used by https://github.com/spacemeshos/go-spacemesh/pull/5061)

## Test Plan
<!-- Please specify how these changes were tested
(e.g. unit tests, manual testing, etc.) -->

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed
- [x] Update [changelog](../CHANGELOG.md) as needed