Remove ctx field from Generator and Certifier (#5180)
## Motivation
It's considered an anti-pattern to store a `Context` in a struct. It leads to code that is difficult to reason about.

## Changes
Removed `ctx` field from `Generator` and `Certifier` structs. Instead, their `Start()` methods take a context. The `Stop()` is handled by cancelling the context derived from the context passed to `Start()`.

## Test Plan
Existing tests should pass