in memory index with atx data for consensus layer (#5013)
closes: https://github.com/spacemeshos/go-spacemesh/issues/4164

this change introduces index with minimal data about identities which is required for consensus. the goal is to have predictable latency on the codepath that are sensitive to it, such as proposals and even more hare messages. the number of issues with that will only grow when sql db size won't fit in page cache (so querying vrf nonce and atx data will hit the disk often).

cached data per identity:
- weight
- height and base height
- vrf nonce
- malicious boolean

the data is added to the cache either on startup or when new atx is received. nonce may be loaded from previous atx if it wasn't included in this one. when block from epoch X is applied, we will evict epochs starting from X-capacity. index is configured to fit same number of epochs that fit in the tortoise window, this enables to self-healing during this period.

internally we store data indexed by epoch for simple eviction, and each epoch has an index by atx.

memory estimate for 1_000_000 identities is ~140MB