remove hare1 code (#5256)
on mainnet we are running certificate with 200 committee size, while hare3 runs with 400.
to keep config compatible i added a separate configuration variable, it defaults to hare3 committee size on every network except mainnet. and on mainnet it will require scheduled upgrade to increase comittee size.

the rest of the change is removal of hare1 code and some dependencies that were spread around the codebase.