Fix flaky test in timesync package (#5303)
## Motivation
Fixes a flaky test in the `timesync` package.

## Changes
`Test_NodeClock_Await_BeforeGenesis` checks that awaiting a layer happens timely (within one tick interval). The test occasionally fails, when the tick interval aligns with the layer timing causing it to be a few µs late. The fix is to increase the allowed "lateness" slightly to not fail when this happens.

## Test Plan
n/a

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed
- [x] Update [changelog](../CHANGELOG.md) as needed