migration to drop activesets with vacuum (#5406)
the change is to get rid of the bloated (and not so useful) state by executing:

```
delete from activesets where epoch < 12;
vacuum;
pragma wal_checkpoint(TRUNCATE);
```

#### main.db-vacuum-state 

we want to execute vacuum only once, for that i added a configuration parameter:

main.db-vacuum-state = 8

the example above specifies if database version is below or equal to 8 we will run vacuum after applying recent migration.
once migration 9 is applied we will not run vacuum again, unless main.db-vacuum-state was changed.

#### main.db-skip-migrations

we also want to preserve all activesets on single node, just to be extra-safe. so that if we for some reason in the future will need them, we can recover them from such archive node.

main.db-skip-migrations = [9]

it accepts list, if migration is in the list it won't be executed. instead we will simply bump the sqlite user version.
migration 9 is where activesets are deleted. note that if this parameter is set, it is better to disable main.db-vacuum-state by setting it to zero.