Fix TestSyncSimulateMultiple (#5430)
## Motivation
The test was flaky because the instances were created in a loop and started immediately. The test would fail if any goroutine called in `sync.Start()` started before other peers registered in the protocol.

## Changes
Create all peers first (so they register in the protocol) and then start them.

## Test Plan
Passed 10k test runs.