fetch: add streaming mode
In streaming mode, no extra buffering is used for fetch server and
client. Fetch client still uses some big slices such as list of ATX
IDs in an epoch, but that's to be improved in a followup.

Streaming mode is enabled via `fetch.streaming` boolean config key,
which defaults to false.