sync: disable mesh agreement on mainnet (#5522)
the goal of mesh agreement is to download ballots and activations when partitions made individual progress.
such as in Partition_50_50 or 70_30 tests.

however on mainnet it creates unnecessary load by triggering queries to download atx. i suggest to disable that and wait for set reconciliation to fix it.

related: https://github.com/spacemeshos/go-spacemesh/issues/5519