hare/eligibility: switch to using atxsdata for preparing activeset
hare oracle needs to maintain weights for active identitities.
before this change it was using lru cache to create an internal mapping
from node id to weight.

with this change we get that data from atxsdata, which keeps atxs for atleast two epochs
and more compact. note that we technically could store a pointer to atx id which will be stored
in atxdata, but for simplicity we store tuple with atx id and weight in hare oracle.

additionally i cleaned codebase in hare/eligibility a bit to make modifying it easier.