miner: use atxsdata to generate activeset and persist it
this change removes lru cache from miner module.
instead it uses atxsdata, that stores atx header in a more compact way.
beside that it generates atx grades by iterating over timestamps stored in database
and persists prepared activeset, so that this work is not repeated everytime when
node is restarted.