Adjust to 2.2M ATXs (#5571)
## Motivation

Current limits allow for 1.5M ATXs. This leaves room for a ~36% increase over the previous epoch. Growth in the current epoch may exceed this, so prepare for that chance and increase the limits.

(I realise that this change _may_ be obsolete with upcoming changes to sync. It may be helpful to merge and backport this change to a 1.3.x release early anyway, just in the sync changes either won't obsolete the ATX limit increase or the sync changes don't become ready in time for the next CG.)