metrics: meter how long request spends in queue and sql connection wait time (#5588)
introduced two metrics:
- histogram p2p_server_in_queue_latency_seconds 

describes how long each request waited in the queue before it was received by the handler, for example if queue is overly long, such as size=1000 and req/s=10, request may remain in queue for 100 intervals where interval is usually 1s, which is probably a misconfiguration and will lead to many timed out requests.

- histogram p2p_conn_wait_seconds

number of connections to sqlite is configured in advance and is a way to limit memory usage. however if requests are taking too long then the likelihood of starvation is high, and this is a metric to monitor that. generally should remain very low