revert window size to 10k (#5627)
closes: https://github.com/spacemeshos/go-spacemesh/issues/5623

this value needs to be versioned in order for sync to work reliably. it has to be 10 000 when 28000 layers are synced, otherwise tortoise can't decode many ballots, which breaks the sync completely.

it can be changed to lower number when layers 28000++ are processed, but some code to make it versioned is not so trivial.