download atxs from current epoch in background without blockinng syncedness (#5600)
closes: https://github.com/spacemeshos/go-spacemesh/issues/5553

this change is on top of https://github.com/spacemeshos/go-spacemesh/pull/5599, and will be rebased after that one is merged

the change always spawns a background worker to download atxs for the ongoing epoch. such background worker for epoch X is spawned when half of the layers of epoch X have passed. this is done so that we always do some useful work to do. such worker will be asking N peers (with a default of 2) every 30 minutes, until we are ready to spawn worker for epoch X + 1.

the side effect of this change, is that we are not going to block syncedness if node has all previous atx, but not from the last epoch. so node will be able to rejoin consensus faster after restart without any risk.